Significant changes and tidy up
Been a while since the last commit. I've tidied up the file structure in to a similar system as for exopt.

There's offline code and online code. I still haven't got the online code in yet because I access the code via ssh so I'll need to figure out how to do that.

The focus has been on PID control. There's a nice visualisation for the offline code (showing you reference vs. , but I haven't been able to figure out how to get it to work for online, although that shouldn't matter too much when I start working with the exoskeleton (soon).