Final results file
Accidentally changed kappas so it's not actually in steps of 0.1! But this doesn't matter too much, changed it back though.

Have saved the final results in optimisation-results-final.mat, along with storing the known maxima, and arrays for easy plotting using:

figure
surf(x,y,thigh_x_mean.')
view(2)

For example.